cambio en la sintaxis de number en productos, el error the query is empty sigue presente al iniciar el servidor

+++ productos.js
@@ -16,7 +16,7 @@ class Productos {
                         table.increments('id')
                         table.string('tittle')
                         table.string('descripcion')
-                        table.float('price')
+                        table.integer('price')
                         table.string('foto_url')
                     } else {
                         table.increments('id')
